[!][update] move `hikami_core::trap` module to `hikami`
