+++ hikami_core/Cargo.toml
@@ -4,7 +4,6 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
-extension_manager = { path = "../extension_manager" }
 elf = { workspace = true }
 fdt = { workspace = true }
 raki = { workspace = true }
