+++ src/trap/exception.rs
@@ -5,19 +5,19 @@ mod page_fault_handler;
 mod sbi_handler;
 
 use super::hstrap_exit;
-use crate::guest;
-use crate::h_extension::{
+use hikami_core::guest;
+use hikami_core::h_extension::{
     csrs::{htval, vstvec},
     HvException,
 };
-use crate::HYPERVISOR_DATA;
-use sbi_handler::sbi_call;
+use hikami_core::HYPERVISOR_DATA;
 
 use core::arch::asm;
 use riscv::register::{
     scause::{self, Exception},
     stval,
 };
+use sbi_handler::sbi_call;
 use sbi_handler::{
     sbi_base_handler, sbi_fwft_handler, sbi_pmu_handler, sbi_rfnc_handler, sbi_time_handler,
 };
