+++ src/trap/exception/sbi_handler.rs
@@ -1,7 +1,7 @@
 //! Handle VS-mode Ecall exception  
 //! See [https://github.com/riscv-non-isa/riscv-sbi-doc/releases/download/v2.0/riscv-sbi.pdf](https://github.com/riscv-non-isa/riscv-sbi-doc/releases/download/v2.0/riscv-sbi.pdf)
 
-use crate::h_extension::csrs::{hvip, VsInterruptKind};
+use hikami_core::h_extension::csrs::{hvip, VsInterruptKind};
 
 use riscv::register::sie;
 use sbi_rt::SbiRet;
